Break import cycle in reports by merging files
