+++ packages/lesswrong/lib/collections/reports/collection.js
@@ -1,5 +1,5 @@
 import schema from './schema.js';
-import './permissions.js'
+import Users from 'meteor/vulcan:users';
 import { createCollection, getDefaultResolvers, getDefaultMutations} from 'meteor/vulcan:core';
 import { addUniversalFields } from '../../collectionUtils'
 
@@ -12,5 +12,26 @@ const Reports = createCollection({
 });
 
 addUniversalFields({collection: Reports})
+
+const membersActions = [
+  'reports.new',
+  'reports.view.own',
+];
+Users.groups.members.can(membersActions);
+
+const sunshineRegimentActions = [
+  'reports.new',
+  'reports.edit.all',
+  'reports.remove.all',
+  'reports.view.all',
+];
+Users.groups.sunshineRegiment.can(sunshineRegimentActions);
+
+Reports.checkAccess = (user, document) => {
+  if (!user || !document) return false;
+  return (
+    document.userId === user._id ? Users.canDo(user, 'reports.view.own') : Users.canDo(user, `reports.view.all`)
+  )
+};
 
 export default Reports;
